Delete the uController framework
