+++ samples/Sample.csproj
@@ -8,7 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\uController\uController.csproj" />
     <ProjectReference Include="..\src\uController.SourceGenerator\uController.SourceGenerator.csproj" OutputItemType="Analyzer" ReferenceOutputAssembly="false" />
   </ItemGroup>
