+++ src/uController.SourceGenerator/CodeGenerationTypes.cs
@@ -3,38 +3,6 @@
 using System;
 using System.Runtime.CompilerServices;
 
-namespace uController
-{
-    [TypeForwardedFrom("uController")]
-    public class HttpHandler { }
-    [TypeForwardedFrom("uController")]
-    public class IHttpRequestReader { }
-    [TypeForwardedFrom("uController")]
-    public class ObjectResult { }
-    [TypeForwardedFrom("uController")]
-    public class RouteAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class HttpMethodAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromQueryAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromHeaderAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromBodyAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromFormAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromRouteAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromCookieAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class FromServicesAttribute { }
-    [TypeForwardedFrom("uController")]
-    public class JsonRequestReader { }
-    [TypeForwardedFrom("uController")]
-    public class JsonResponseWriter { }
-}
-
 namespace Microsoft.AspNetCore.Http
 {
     [TypeForwardedFrom("Microsoft.AspNetCore.Http.Abstractions")]
