+++ src/uController.SourceGenerator/uController.SourceGenerator.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
@@ -13,9 +13,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="..\uController\AwaitableInfo.cs" Link="AwaitableInfo.cs" />
-    <Compile Include="..\uController\HttpModel.cs" Link="HttpModel.cs" />
-    <Compile Include="..\uController\ReflectionExtensions.cs" Link="ReflectionExtensions.cs" />
 
     <None Include="$(OutputPath)\$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
   </ItemGroup>
