--- src/uController/EndpointRouteBuilderExtensions.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Routing;
-using uController;
-
-namespace Microsoft.AspNetCore.Builder
-{
-    public static class EndpointRouteBuilderExtensions
-    {
-        public static void MapHttpHandler<THttpHandler>(this IEndpointRouteBuilder builder)
-        {
-            HttpHandlerBuilder.Build<THttpHandler>(builder);
-        }
-    }
-}
