--- src/uController/FromBody.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace uController
-{
-    [AttributeUsage(AttributeTargets.All, Inherited = false, AllowMultiple = false)]
-    public sealed class FromBodyAttribute : Attribute
-    {
-    }
-}
