--- src/uController/FromRoute.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace uController
-{
-    [AttributeUsage(AttributeTargets.All, Inherited = false, AllowMultiple = false)]
-    public sealed class FromRouteAttribute : Attribute
-    {
-        public string Name { get; private set; }
-
-        public FromRouteAttribute()
-        {
-        }
-
-        public FromRouteAttribute(string name)
-        {
-            Name = name;
-        }
-    }
-}
