--- src/uController/IResult.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-
-namespace uController
-{
-    // Result is the interface version of a RequestDelegate
-    public interface IResult
-    {
-        Task ExecuteAsync(HttpContext httpContext);
-    }
-}
