--- src/uController/Results.cs
@@ -1,14 +1,0 @@
-﻿using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-
-namespace uController
-{
-    public static class Results
-    {
-        public static IResult BadRequest() => Status(StatusCodes.Status400BadRequest);
-        public static IResult NotFound() => Status(StatusCodes.Status404NotFound);
-        public static IResult Ok() => Status(StatusCodes.Status200OK);
-        public static IResult Ok(object value) => new ObjectResult(value);
-        public static IResult Status(int statusCode) => new StatusCodeResult(statusCode);
-    }
-}
