--- src/uController/StatusCodeResult.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-
-namespace uController
-{
-    public class StatusCodeResult : IResult
-    {
-        public int StatusCode { get; }
-
-        public StatusCodeResult(int statusCode)
-        {
-            StatusCode = statusCode;
-        }
-
-        public Task ExecuteAsync(HttpContext httpContext)
-        {
-            httpContext.Response.StatusCode = StatusCode;
-            return Task.CompletedTask;
-        }
-    }
-}
