+++ uController.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.4.32728.343
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sample", "samples\Sample.csproj", "{E75EECDA-E740-454B-BA83-7B7AC0D9E0A9}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "uController", "src\uController\uController.csproj", "{57549041-6101-4D98-AA83-E4856CCBCA71}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "uController.SourceGenerator", "src\uController.SourceGenerator\uController.SourceGenerator.csproj", "{045C64E7-5639-4ECD-BEE3-57597E7CF77E}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{9AC9FBC2-655D-4D1C-B3B0-8429140C1125}"
@@ -29,10 +27,6 @@ Global
 		{E75EECDA-E740-454B-BA83-7B7AC0D9E0A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E75EECDA-E740-454B-BA83-7B7AC0D9E0A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E75EECDA-E740-454B-BA83-7B7AC0D9E0A9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{57549041-6101-4D98-AA83-E4856CCBCA71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{57549041-6101-4D98-AA83-E4856CCBCA71}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{57549041-6101-4D98-AA83-E4856CCBCA71}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{57549041-6101-4D98-AA83-E4856CCBCA71}.Release|Any CPU.Build.0 = Release|Any CPU
 		{045C64E7-5639-4ECD-BEE3-57597E7CF77E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{045C64E7-5639-4ECD-BEE3-57597E7CF77E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{045C64E7-5639-4ECD-BEE3-57597E7CF77E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -43,7 +37,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{E75EECDA-E740-454B-BA83-7B7AC0D9E0A9} = {37EABDC7-EC04-40DB-AADB-73232A885C63}
-		{57549041-6101-4D98-AA83-E4856CCBCA71} = {9AC9FBC2-655D-4D1C-B3B0-8429140C1125}
 		{045C64E7-5639-4ECD-BEE3-57597E7CF77E} = {9AC9FBC2-655D-4D1C-B3B0-8429140C1125}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
